Improvements to TAP library add-ons for Kerberos and remctl
Add a TAP test library Kerberos initialization function that uses
kinit rather than native Kerberos functions in addition to the native
version.

Add TAP test library functions to clean up the Kerberos ticket cache
and to cleanly shut down remctld.

Add TAP shell libraries for initializing Kerberos and starting and
stopping remctld.

Don't require Kerberos to build this package, but include the sample
probes commented-out in configure.ac.

+++ tests/tap/remctl.c
@@ -36,7 +36,7 @@ remctld_start(const char *remctld, const char *principal, const char *config)
     struct timeval tv;
     size_t n;
 
-    pidfile = concatpath(getenv("BUILD"), "data/remctl.pid");
+    pidfile = concatpath(getenv("BUILD"), "data/remctld.pid");
     if (access(pidfile, F_OK) == 0)
         if (unlink(pidfile) != 0)
             sysbail("cannot delete %s", pidfile);
@@ -71,10 +71,16 @@ void
 remctld_stop(pid_t child)
 {
     char *pidfile;
+    struct timeval tv;
 
-    kill(child, SIGTERM);
-    waitpid(child, NULL, 0);
-    pidfile = concatpath(getenv("BUILD"), "data/remctl.pid");
+    tv.tv_sec = 0;
+    tv.tv_usec = 10000;
+    select(0, NULL, NULL, NULL, &tv);
+    if (waitpid(child, NULL, WNOHANG) == 0) {
+        kill(child, SIGTERM);
+        waitpid(child, NULL, 0);
+    }
+    pidfile = concatpath(getenv("BUILD"), "data/remctld.pid");
     unlink(pidfile);
     free(pidfile);
 }
